context: Use proper status return codes
This will allow us to restrict our logic for "should i do a thing" to
the functions instead of trying to bloat the struct with every possible
condition. Such a method would be required for the systemd units, and
ensuring we never reload these even in a chroot.

Another example would be running clr-boot-manager update, we only want
to do this when in a chroot, *unless* we're being forced to. Even under
a forced mode it would be stupid to reload systemd (we cannot know if it
will ever work or be available) but cbm is something we would want to support.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>