util: Add new `usc_get_strn_component` function
We'll need this to help us expand the depmod code to handle subtrees.
Effectively this walks the string to the requested depth and returns
a duplicated string for that indexed path component.

We'll use this to look at any given path and extract the component from
it that we want, which will be used by the depmode code to determine
the kernel tree's identifier to pass to /sbin/depmod.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>