handlers/depmod: Learn about kernel tree subpaths
This allows us to not only spot new kernels, but to spot kernel modules
that have been injected into the kernel tree from drivers package. We'll
look within the root of the tree (i.e. `./extra`) and within the location
we'd expect nvidia to install, `./kernel/drivers/*`.

This allows us to easily track modifications to the tree and reliably record
changes, whilst using the new skip API to not do unnecessary work.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>