handlers/icon-cache: Add matching for hicolor/gnome
This allows us to perform matches on directories modified by packages
shipping their own icon assets within the hicolor and gnome directories
to expose them as part of the icon theme.

With the new granular handling we can collapse to the relevant tree name
and update from there, allowing icon caches to always be correctly built.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>