handlers: Implement a new gconf2 handler
This is a bit of a cheeky handler, in that it double globs. First up
we'll nuke the existing gconf tree, then we'll set about rebuilding it
from the two fixes paths.

This solves many chicken egg issues by ensuring we just rebuild in one
ago, but still get triggered when one of our source directories changes.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>