handlers: Add new handler for `vboxdrv.service`
This is a rather specialist case, if we're not in a chroot and we've
detected an update (or initial install thereof) VirtualBox - we need
to restart the systemd service to ensure the modules are also upgraded.
This ensures VirtualBox remains usable without requiring a reboot, or
is indeed available the first time it is installed.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>