Add the new `udev-rules` trigger
This trigge will reload the udev rules and then apply them via `trigger`
when the source directories change. We run this last as it can cause some
userspace changes/responses that will require the system to be in a fully
consistent state.

This will help in cases like `libgphoto2` update by ensuring that the fixes
to the new udev rules will apply automatically and not require the users to
have to reboot.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>